requirements.txt
streamlit
pandas
google-generativeai
protobuf
grpcio
python-dotenv